requirements.txt
bs4
pyyaml
requests
selenium
webdriver_manager